nagios-4.3.4-cgi_refresh.patch - CGI refresh when Referer data is demanded by the server
Description:

The output from several of the standard Nagios CGI programs is auto-refreshed by means of the HTTP Refresh header. But that mechanism fails to work when we have the Apache web server set up to demand Referer information, which is dropped in such an automated refresh. We patch these CGI programs to instead use a tiny fragment of JavaScript to carry out the page reload at the refresh time.

Status:

	This patch applies exactly.  It modifies Nagios CGI programs so
	they operate smoothly in the GroundWork context.