nagios-4.3.4-Makefile.in.statuswml2.patch - statuswml support for WAP2
Description:

Provide an alternate CGI binary for customers wishing to deploy WAP 2.0 (used on modern phones) instead of WAP 1.1. All this takes is a copy of statuswml.c renamed to statuswml2.c and with the DTD for WAP (WML) 1.1 changed to the corresponding DTD for WAP 2.0. With appropriate instructions for putting the resultant statuswml2.cgi binary into operation, the customer will be responsible for setting it up on-site.

Status:

	This patch applies exactly.  It requires the companion
	cgi/statuswml2.c program specific to this Nagios release.