nagios-4.3.4-website_url.patch - support the website_url directive in the GroundWork context
Description:

Nagios 4.3.4 introduces the website_url directive in the main configuration file, to specify the base URL used to dynamically construct the values of the new $HOSTINFOURL$ and $SERVICEINFOURL$ macros. However, the upstream code assumes that the final URL for these macros will include a /cgi-bin path component immediately after the website_url value substituted into the complete URL for the macro value. That does not comport with how we make Nagios CGI pages available in the GroundWork portal. So we edit the Nagios code to remove the fixed /cgi-bin path component from these macro expansions, allowing us to specify a value for the website_url directive that does not require a path rewrite in the Apache httpd.conf file.

Status:

	This patch applies exactly.  It modifies the pathnames used within
	Nagios to create values for the $HOSTINFOURL$ and $SERVICEINFOURL$
	macros, to reflect corresponding pathnames as they apply in the
	GroundWork context.