nagios-4.3.4-traceroute_cgi_perl.patch - port traceroute.cgi to GWMEE
Description:

The traceroute.cgi program provided in the Nagios distribution seems never to have been ported to the GroundWork context, in previous releases of Nagios and GWMEE. We do so now with this patch. This probably won't have much significant impact; it's being done for completeness, so we can test this CGI program to see if it's useful to us.

Status:

	This patch applies exactly.  It converts contrib/traceroute.cgi
	to live in the GroundWork Monitor context.