GROUNDWORK-1675-result-statistics:  extend Nagios to provide queue stats
Modify our existing patches to Nagios that insert and remove items
from the check-result queue we use to pass data from Bronx to Nagios.
The effect is to efficiently keep track of how many items are in the
queue at any time.  Provide a means to safely retrieve that info and
some data about the items at either end of the queue, so an outside
agency can have some degree of observability as to processing latency.