GROUNDWORK-1675-throughput-stats:  capture data for queue ingress/egress stats
Previously captured statistics allow us to track latency in the Nagios
check results list.  Now we also provide the means to track the rates at
which items are being added and removed from that same list.  This will
provide a measure of actual throughput when the list size seems to be
fairly steady.