GROUNDWORK-1719-fix-passive-check-logging:  shuffle passive-check logging
The upstream Nagios 4.4.0 through 4.4.6 releases (at least) contain
a refactoring bug that causes inappropriate data to be logged in
PASSIVE HOST CHECK and PASSIVE SERVICE CHECK messages that appear in
the nagios.log file.  The changes in this check-in fix that by unwinding
the broken refactoring, and restoring the affected log messages to where
they need to be situated so they display the expected new data.

If these fixes get accepted upstream and we later upgrade to some Nagios
base release from 4.4.7 onward, it is quite likely that the rebasing of
our GROUNDWORK branch on the new upstream code will result in a silent
broken merge of the upstream fix and our own fix.  As with all changes
that GroundWork applies, we will need to pay close attention to every
one of the merge points to ensure that it got done correctly.  But in
this case, since we are recommending that the upstream developers adopt
our fix, I would be very surprised if automated git merge doesn't break
the code.  We are forewarned.