GROUNDWORK-1718-add-broker-calls:  extend Nagios support for DataGeyser
Add support within Nagios itself to provide a means for DataGeyser to see
the states of individual hosts and services right before check results
are applied to those objects.  DataGeyser can then cache that early
state and compare it to the state after the check results are applied,
to see which data items actually need to be sent downstream.

Also fix a few Nagios bugs we spotted while doing the analysis for the
changes related to DataGeyser.  These bugs relate to how downtime and
latency data is handled.  This bug fixes have been noted, and we will
eventually report them upstream (though the Nagios folks don't seem to
be paying attention to any bug reports in the last couple of years.)